services/aicut-orchestrator/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aicut"
version = "0.1.0"
description = "Constraint-based multi-asset auto editing pipeline stitched from existing open-source tools."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "faster-whisper>=1.1.0",
  "opentimelineio>=0.17.0",
  "pydantic>=2.8.0",
  "pyyaml>=6.0.2",
  "rich>=13.7.1",
  "scenedetect[opencv]>=0.6.7",
  "typer>=0.12.3",
]

[project.optional-dependencies]
diarization = [
  "pyannote.audio>=4.0.0",
  "torch>=2.3.0",
]
dev = [
  "pytest>=8.2.0",
]

[project.scripts]
aicut = "aicut.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
